# lib/api-spec/openapi.orval.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Api",
    "version": "0.2.0",
    "description": "Docuvia — Universal VCS Knowledge Graph Engine API"
  },
  "servers": [{ "url": "/api", "description": "Base API path" }],
...
  "paths": {
...
    "/projects/{id}/l2-nodes": {
      "get": {
        "operationId": "listProjectL2Nodes",
        "tags": ["l2-nodes"],
        "summary": "List L2 nodes for a project",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "List of L2 nodes",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/L2Node" } }
              }
            }
          }
        }
      }
    },
    "/projects/{id}/l2-nodes/confirm-bootstrap": {
      "post": {
        "operationId": "confirmBootstrap",
        "tags": ["l2-nodes"],
        "summary": "Confirms the AI-discovered L2 modules, saves glob path patterns, and switches the project out of bootstrap mode.",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["approvedModules", "rejectedModuleIds"],
                "properties": {
                  "approvedModules": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": ["id", "pathPatterns"],
                      "properties": {
                        "id": { "type": "integer" },
                        "pathPatterns": { "type": "array", "items": { "type": "string" } }
                      }
                    }
                  },
                  "rejectedModuleIds": { "type": "array", "items": { "type": "integer" } }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bootstrap confirmed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["success", "message"],
                  "properties": {
                    "success": { "type": "boolean" },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/l2-nodes": {
      "post": {
        "operationId": "createL2Node",
        "tags": ["l2-nodes"],
        "summary": "Create a new L2 node",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/L2NodeInput" } }
          }
        },
        "responses": {
          "201": {
            "description": "Created L2 node",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/L2Node" } }
            }
          }
        }
      }
    },
    "/l2-nodes/{id}": {
      "patch": {
        "operationId": "updateL2Node",
        "tags": ["l2-nodes"],
        "summary": "Update an L2 node",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/L2NodeUpdate" } }
          }
        },
        "responses": {
          "200": {
            "description": "Updated L2 node",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/L2Node" } }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteL2Node",
        "tags": ["l2-nodes"],
        "summary": "Delete an L2 node",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": { "204": { "description": "Deleted" } }
      }
    },
    "/l2-nodes/{id}/links": {
      "get": {
        "operationId": "listNodeLinks",
        "tags": ["l2-nodes"],
        "summary": "Get dependency links for an L2 node",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "List of node links",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/NodeLink" } }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createNodeLink",
        "tags": ["l2-nodes"],
        "summary": "Create a dependency link between L2 nodes",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/NodeLinkInput" } }
          }
        },
        "responses": {
          "201": {
            "description": "Created link",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/NodeLink" } }
            }
          }
        }
      }
    },
    "/l2-nodes/{id}/l3-nodes": {
      "get": {
        "operationId": "listL3Nodes",
        "tags": ["l3-nodes"],
        "summary": "List L3 nodes under an L2 node",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "List of L3 nodes",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/L3Node" } }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createL3Node",
        "tags": ["l3-nodes"],
        "summary": "Create a new L3 node",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/L3NodeInput" } }
          }
        },
        "responses": {
          "201": {
            "description": "Created L3 node",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/L3Node" } }
            }
          }
        }
      }
    },
    "/l3-nodes/{id}": {
      "patch": {
        "operationId": "updateL3Node",
        "tags": ["l3-nodes"],
        "summary": "Update an L3 node",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/L3NodeUpdate" } }
          }
        },
        "responses": {
          "200": {
            "description": "Updated L3 node",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/L3Node" } }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteL3Node",
        "tags": ["l3-nodes"],
        "summary": "Delete an L3 node",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": { "204": { "description": "Deleted" } }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "L2Node": {
        "type": "object",
        "required": [
          "id",
          "projectId",
          "name",
          "type",
          "l3Count",
          "l1TagIds",
          "aiGenerated",
          "createdAt"
        ],
        "properties": {
          "id": { "type": "integer" },
          "projectId": { "type": "integer" },
          "name": { "type": "string" },
          "type": { "type": "string", "enum": ["package", "module", "pcd", "sys-uncategorized"] },
          "description": { "type": "string", "nullable": true },
          "l3Count": { "type": "integer" },
          "l1TagIds": { "type": "array", "items": { "type": "integer" } },
          "aiGenerated": { "type": "boolean" },
          "needsReview": { "type": "boolean" },
          "isSystem": { "type": "boolean" },
          "isBootstrapConfirmed": { "type": "boolean" },
          "pathPatterns": { "type": "array", "items": { "type": "string" } },
          "createdAt": { "type": "string" }
        }
      },
      "L2NodeInput": {
        "type": "object",
        "required": ["projectId", "name", "type"],
        "properties": {
          "projectId": { "type": "integer" },
          "name": { "type": "string", "minLength": 1 },
          "type": { "type": "string", "enum": ["package", "module", "pcd", "sys-uncategorized"] },
          "description": { "type": "string" },
          "l1TagIds": { "type": "array", "items": { "type": "integer" } },
          "aiGenerated": { "type": "boolean" }
        }
      },
      "L2NodeUpdate": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "minLength": 1 },
          "type": { "type": "string", "enum": ["package", "module", "pcd", "sys-uncategorized"] },
          "description": { "type": "string" },
          "l1TagIds": { "type": "array", "items": { "type": "integer" } },
          "needsReview": { "type": "boolean" }
        }
      },
      "NodeLink": {
        "type": "object",
        "required": ["id", "sourceNodeId", "targetNodeId", "linkType", "createdAt"],
        "properties": {
          "id": { "type": "integer" },
          "sourceNodeId": { "type": "integer" },
          "targetNodeId": { "type": "integer" },
          "linkType": { "type": "string" },
          "sourceNodeName": { "type": "string", "nullable": true },
          "targetNodeName": { "type": "string", "nullable": true },
          "createdAt": { "type": "string" }
        }
      },
      "NodeLinkInput": {
        "type": "object",
        "required": ["targetNodeId"],
        "properties": { "targetNodeId": { "type": "integer" }, "linkType": { "type": "string" } }
      },
      "L3Node": {
        "type": "object",
        "required": ["id", "l2NodeId", "title", "nodeType", "aiGenerated", "createdAt"],
        "properties": {
          "id": { "type": "integer" },
          "l2NodeId": { "type": "integer" },
          "title": { "type": "string" },
          "content": { "type": "string", "nullable": true },
          "nodeType": { "type": "string", "enum": ["change", "rule", "decision", "context"] },
          "commitHash": { "type": "string", "nullable": true },
          "aiGenerated": { "type": "boolean" },
          "confidence": { "type": "number", "nullable": true },
          "noiseScore": { "type": "number", "nullable": true },
          "createdAt": { "type": "string" }
        }
      },
      "L3NodeInput": {
        "type": "object",
        "required": ["l2NodeId", "title", "nodeType"],
        "properties": {
          "l2NodeId": { "type": "integer" },
          "title": { "type": "string", "minLength": 1 },
          "content": { "type": "string" },
          "nodeType": { "type": "string", "enum": ["change", "rule", "decision", "context"] },
          "commitHash": { "type": "string" },
          "aiGenerated": { "type": "boolean" },
          "confidence": { "type": "number" }
        }
      },
      "L3NodeUpdate": {
        "type": "object",
        "properties": {
          "title": { "type": "string", "minLength": 1 },
          "content": { "type": "string" },
          "nodeType": { "type": "string", "enum": ["change", "rule", "decision", "context"] },
          "noiseScore": { "type": "number" }
        }
      },
...
    }
  }
}